Fixed some issues with the IRIX/OpenSceneGraph configuration:
MIPSPro needs a -LANG:std option to properly compile STL stuff (which the
OSG headers require).  Added configure.ac code to check the SGI compiler
and add the -LANG:std option if the compiler is version 7.3 or later.

Added -n32 -mips3 to the CFLAGS under IRIX.  These were missing and were
causing the autoconf library checks to fail.

Added /usr/freeware/include to the include path for freetype.  For some
reason ft2build.h is not in the directory with all the other freetype
headers (/usr/freeware/include/freetype2/freetype).