Changed vsSequencer to use a vsTimer for timing so that it truly is a common class (gettimeofday() does not exist in Windows).
Fixed a bug in the while and if statements in update (the time comparisons
were being done wrong).

Since usleep() typically only has a resolution of 0.01 seconds (the
same as the process scheduler), added code to use a busy-wait loop if the
amount of time to wait is less than 0.01 seconds.  Otherwise, the wait time
will always be at least 0.01 seconds.

Added a check before the usleep() to avoid a negative float value being
turned into a very large unsigned long value and then passed to usleep().