This class allows VESS programmers to use multiple Performer draw callbacks. It has several methods to make this feature easy to use. These methods are split into two classes, application methods and callback methods.
Application methods: These methods are called on the list object itself.
The prepend and append callback methods take a function to be added to the
list and either a pointer indicating the size of the shared memory to be
allocated for that callback node, or a pointer to shared memory already
allocated. The latter of these options can be used in case multiple callbacks
need to work on the same data. A single method for removing callbacks is
provided that takes the function pointer of the callback node and a pointer
to the shared memory on which that node acts.

In the case