Modified vsParticleSystem to use only one vsGeometry for all particles in hardware mode, instead of a vsGeometry existing for each particle.  Software mode is not changed.
Also added new constructors that allow a pre-existing shader or GLSL program
attribute to be specified, instead of specifying the shader code filename.
This allows a single attribute to be shared among many particle systems, and
avoids the cost of a recompile/relink of the shader code for every new
particle system that is created.