Added a much more complete animation API to vsCharacter.  vsCharacter now supports several levels of animation (default, looping, one-time, and final).
Each of the start animation methods also takes a transition time, allowing
the user to specify how long it should take to transition between animations.
Transitions are done via linear interpolation between the current animation
pose and the starting pose of the new animation.

To give proper credit, the ideas and some of the code behind this API came
from Brandon Haber's original work on the DIVAARS animated characters, and
Dustin Chertoff's efforts on the GamePAB software.