Changes to make vsObjectMap thread-safe.  This is needed because any creation or destruction of any vsNode descendant object modifies a vsObjectMap that is a class member of vsNode.
I'm not sure this is the correct solution (should vsObjectMap be thread-safe?).  The alternative would be to put a mutex in the vsNode class and lock it whenever the class's map is modified.  By changing vsObjectMap, we keep the locking self-contained, and the code is cleaner, but one could argue that the problem really isn't vsObjectMap's fault, but rather how it's being used.