Complete re-write of vsVRPNTrackingSystem and vsHiball.
The original implementation missed a key part of the VRPN system, so this
re-write was necessary.  One thing to watch out for with VRPN trackers, is
that there is no explicit way to query for the number of trackers in a
tracking system.  The only way to know is to get updates.  Unfortunately,
systems like the Hiball will only send updates if the tracker is currently
acquired by the system.  The vsVRPNTrackingSystem class will dynamically
add trackers if they come online late, but the application code must be
written to handle this.

In the process, I also rearchitected the class structure to be more in line
with other VESS tracking systems.  Now, the vsVRPNTrackingSystem has all of
the functionality built-in, and the vsHiball is a simple pass-through class
(much like vsFlockOfBirds is to vsAscensionSerialTrackingSystem).  The previous
structure didn't make much sense to me.